requirements.txt
python_version ==3.7
fastapi==0.100.0
uvicorn==0.22.0
pyannote.audio==2.1.1
pandas==1.3.5
python-multipart==0.0.6

